Structure refactoring, splitting to smaller logical files
