+++ setup.py
@@ -29,10 +29,13 @@ setup(
     long_description=read('README.rst', 'CHANGES.rst'),
     entry_points={
         'console_scripts': [
-            'ti = ti:main',
+            'ti = ti.time_tracking:main',
         ]
     },
-    install_requires=["colorama", "pyyaml"],
+    install_requires=[
+        "colorama",
+        "PyYAML",
+    ],
     setup_requires=["pytest-runner"] if testing else [],
     tests_require=["pytest", "cram", "pytest-cram"],
     extras_require={
